+++ pp/samples/autoplacer/coupler.py
@@ -7,11 +7,11 @@ from pp.routing.add_fiber_array import add_fiber_array
 
 
 @pp.cell
-def CP2x2(gap=0.3, length=10.0):
+def coupler2x2(gap=0.3, length=10.0):
     c = coupler(gap=gap, length=length)
-    return add_fiber_array(c)
+    return add_fiber_array(component=c)
 
 
 if __name__ == "__main__":
-    c = CP2x2()
+    c = coupler2x2()
     c.show()